+++ client/include/ui/widget/EditTaskDialog.h
@@ -1,20 +1,20 @@
 #pragma once
-#include <QDialog>
-#include <QLineEdit>
-#include <QComboBox>
+#include "ui/widget/CreateTaskDialog.h"
 #include "DataStructures.h"
 
-class EditTaskDialog : public QDialog {
+class EditTaskDialog : public CreateTaskDialog {
     Q_OBJECT
+
 public:
     explicit EditTaskDialog(const Task& task, QWidget* parent = nullptr);
 
     signals:
         void taskUpdated();
 
+protected slots:
+    void handleSubmit() override;
+
 private:
     int taskId;
-    QLineEdit* titleInput;
-    QLineEdit* tagInput;
-    QComboBox* assigneeCombo;
+    QPushButton* btnDelete;
 };