+++ changelog.md
@@ -1,1 +1,2 @@
-* Update Create Task and group Dialogs too look like their equivalent cards.
+* Prevent Tasks from being created with no name or tag
+* Update Edit task dialog to match create task dialog